requirements.txt
# Phase 1 - Extraction
PyMuPDF==1.25.3
pdfplumber==0.11.4
pytesseract==0.3.13
Pillow==11.1.0
python-docx==1.1.2
pydantic==2.10.6
python-magic==0.4.27

# Phase 2 - Cleaning 
requests==2.32.3